requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Vector database
pgvector==0.2.4

# Machine Learning & NLP
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.36.0
spacy==3.7.2
medspacy==1.0.0
scikit-learn==1.3.2

# Data processing
pandas==2.1.3
numpy==1.25.2
jsonlines==4.0.0

# PHI de-identification
philter-py==1.0.5

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
typer==0.9.0
rich==13.7.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1 